+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # Changelog
 
+### 0.2.2
+- Customization ability for the Material-UI button. [FR5](https://github.com/Aluminati/meteor-react-autoform/issues/5)
+- Allow the use of a custom button instead of our default Material-UI. [FR5](https://github.com/Aluminati/meteor-react-autoform/issues/5)
+
+### 0.2.1
+- `useFields` prop has been added back (removed by mistake)
+
 ### 0.2.0
 - `collection` prop type has been removed
 - `schema` prop type has been added and expects your JSON schema (not simple schema)
 - Automated testing has been added
 
 ### 0.1.1
-- Resolved form className prop. See [PR2](https://github.com/Aluminati/meteor-react-autoform/pull/2)
+- Resolved form className prop. [PR2](https://github.com/Aluminati/meteor-react-autoform/pull/2)
 
 ### 0.1.0
 - React v15 compatibility